+++ crates/cratermaker-components/src/surface/mod.rs
@@ -47,7 +47,7 @@ pub struct LocalSurfaceView<'a> {
     pub face_bearing: Option<ArrayView1<'a, f64>>,
     pub region_radius: Option<f64>,
     pub location: Option<(f64, f64)>,
-    pub crater_id: Option<ArrayView2<'a, u64>>,
+    pub crater_id: Option<ArrayView2<'a, u32>>,
     pub desloped_face_elevation: Option<ArrayView1<'a, f64>>,
 }
